Separate dropdown toggle to different file
